Revert "Modify announcement topic to handle when changing sizes or finishing"

+++ srcCpp/RTIDDSImpl.h
@@ -11,6 +11,7 @@
 #include <map>
 #include "MessagingIF.h"
 #include "perftestSupport.h"
+#include "ndds/ndds_cpp.h"
 #include "PerftestTransport.h"
 
 #define RTIPERFTEST_MAX_PEERS 1024
@@ -129,8 +130,8 @@ class RTIDDSImpl : public IMessaging
     bool         _IsReliable;
     bool         _IsMulticast;
     unsigned int _BatchSize;
-    unsigned long _InstanceCount;
-    unsigned long _InstanceMaxCountReader;
+    int          _InstanceCount;
+    int          _InstanceMaxCountReader;
     int          _InstanceHashBuckets;
     int          _Durability;
     bool         _DirectCommunication;
@@ -148,7 +149,7 @@ class RTIDDSImpl : public IMessaging
     int          _peer_host_count;
     char *       _peer_host[RTIPERFTEST_MAX_PEERS];
     bool         _useCft;
-    long _instancesToBeWritten;
+    int          _instancesToBeWritten;
     unsigned int _CFTRange[2];
 
     PerftestTransport _transport;
